Drop a replayed unfollow from the follows cache, like a live commit
The commit success path removes the unfollowed profile from the current
account's follows cache, but the replay success path only re-stamped the
profile shadow. A follows list fetched while the replayed delete was in
flight (at launch, or retained across a bfcache restore) kept the
profile in people-picker surfaces until its next refetch.

The cache helpers move from profile.ts to profile-follows.ts, next to
the query they operate on, so the replay can share them without an
import cycle.
